fix: type in ref dict

+++ tests/test_stats/test_stats.py
@@ -223,7 +223,7 @@ class TestStats:
             "RMSE": np.float64(9.153915273540871),
             "Standard deviation": np.float64(79.32951386752386),
             "Valid count": np.int64(273000),
-            "Total count": 273000,
+            "Total count": np.int64(273000),
             "Percentage valid points": np.float64(100.0),
         }
         assert res_stats_crop == rast_crop.get_stats()
@@ -244,7 +244,7 @@ class TestStats:
             "RMSE": np.float64(10.38534653788665),
             "Standard deviation": np.float64(62.319986152883956),
             "Valid count": np.int64(211527),
-            "Total count": 524000,
+            "Total count": np.int64(524000),
             "Percentage valid points": np.float64(40.36774809160305),
         }
         assert res_stats_crop_proj == rast_crop_proj.get_stats()
@@ -265,7 +265,7 @@ class TestStats:
             "RMSE": np.float64(8.37853254688833),
             "Standard deviation": np.float64(79.45825061580675),
             "Valid count": np.int64(282802),
-            "Total count": 524000,
+            "Total count": np.int64(524000),
             "Percentage valid points": np.float64(53.96984732824428),
         }
         assert stats_masked_rast == rast.get_stats()
@@ -317,7 +317,7 @@ class TestStats:
             "RMSE": np.float64(9.220541807365446),
             "Standard deviation": np.float64(79.44349437534403),
             "Valid count": np.int64(524000),
-            "Total count": 524000,
+            "Total count": np.int64(524000),
             "Percentage valid points": np.float64(100.0),
         }
         assert rast_stats_pc == rast_pc.get_stats()
@@ -340,7 +340,7 @@ class TestStats:
             "RMSE": np.float64(9.153915273540871),
             "Standard deviation": np.float64(79.32951386752386),
             "Valid count": np.int64(273000),
-            "Total count": 273000,
+            "Total count": np.int64(273000),
             "Percentage valid points": np.float64(100.0),
         }
         assert rast_stats_crop_pc == rast_crop_pc.get_stats()
@@ -362,7 +362,7 @@ class TestStats:
             "RMSE": np.float64(10.38534653788665),
             "Standard deviation": np.float64(62.319986152883956),
             "Valid count": np.int64(211527),
-            "Total count": 211527,
+            "Total count": np.int64(211527),
             "Percentage valid points": np.float64(100.0),
         }
         assert rast_stats_crop_proj_pc == rast_crop_proj_pc.get_stats()
